Update patch set 1
Patch Set 1:

After testing this, and thinking a bit more about it - I'm not 100% sure of the value of the cascade=false feature "as it stands".

If in our tile.xml, we declare cascade="false" ( and it only works in the plugin configuration, not the top level <tiles> element AFAIK ) then we don't process the tile - essentially treating it as tho its commented out. It will NEVER be included - that feels wrong.

Is it not that we want to run that tile, but not cascade further into its embedded tiles? Which itself seems wrong, I'm at a loss now to think why we want it. I know we want some form of circular tile prevention - which requires tracking which tiles have already been applied and not repeating oneself. But that's not what this currently does...

Mmmm

Patch-set: 1
Reviewer: Mark Derricutt <1000264@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0