Remove the replaced inheritance assemblers
After doing some more testing - when compleatly removing the tiles
plugin from projects that include an <scm> section only in the parent,
the same behaviour of a broken url/connection/developerConnection appears
in the -effective pom-.

For this reason, I think there's little point in us changing this, and
introducing more code than is neccessary for tiles to operate effectively.

Fixes #6.

Change-Id: I166ebdd11aa82d99fc21f9ff433afd34246ed7fa