Remove the replaced inheritance assemblers
After doing some more testing - when compleatly removing the tiles
plugin from projects that include an <scm> section only in the parent,
the same behaviour of a broken url/connection/developerConnection appears
in the -effective pom-.

For this reason, I think there's little point in us changing this, and
introducing more code than is neccessary for tiles to operate effectively.

Note for review: if we merge this, we should squash out the previous commit
first.

Change-Id: I22446dcdbdf6b79f78735bacd9202ec5a8fa7785