Update patch set 3
Patch Set 3:

It looks like this, and the follow invoker plugin test have revealed that our lifecycle participant is not being triggered on artifacts inside a reactor, unless it's the top project.

Have raised a discussion thread on the dev@maven list to see if its a Maven issue, or an us issue.

Patch-set: 3
Label: Code-Review=0