Add ability to use version ranges for plugins declared in tiles
This feature reduce the risk of bad plugin version resolution when the
same plugin is used in multiple tiles. With "rigid" plugin version, the
plugin could resolve to an old version which for example don't include
required goals.

Also fail the build if intersection of plugin version ranges from
different tiles is overconstrainted and results in an empty set of
acceptable versions.

Change-Id: Ia027100e1b131e256604b2b43b565b015b8165fd (+1 squashed commits)

Squashed commits:

[3f2d8bc] Resolve tile version range to latest SNAPSHOT only if range bounds is explicitely SNAPSHOT

Change-Id: I8f39ecd5426842c1803c8d7bfbe6709ad4f5dd99