Update patch set 1
Patch Set 1:

This is essentially an implementation of the long standing open ticket Apache Maven ticket for the same thing is it not?

I see that ticket has a PR with recent ( yesterday and today ) activity that's more than likely to be resolved, and that introduces a new VersionRangeResultFilter interface, along with a DefaultVersionRangeResultFilter which has the current behaviour.

I'm not sure I like the idea of having two separate behaviours for range resolution - one for tiles and one for normal maven ranges, if we detect we're running on a release that supports the VersionRangeResultFilter, would it be better to inject and use which ever is configured for the current maven session and provide consistency?

[1] https://issues.apache.org/jira/browse/MNG-3092

Patch-set: 1
Label: Code-Review=0