Optionally apply tiles to specific parent
Instead of always injecting tiles after the current project's model
(before the first parent), in build where the tiles are actually defined
in a parent pom we want to apply them before a specific parent. This
allows this specific parent (which probably included the tiles in the
first place) to also overwrite properties from those tiles.

Change-Id: I968ea60a68104a2a9e96b128c6ea7a9ee80dad5c