Update patch set 1
Patch Set 1:

I haven't added it anywhere else no, other than in the Github issue that spawned the update.  I don't think we need to repeat it in the README.adoc, but I also don't think we link to the CONTRIBUTING.adoc from the README.adoc either.

Patch-set: 1
Label: Code-Review=0