Update patch set 3
Patch Set 3:

yes, it does work and we heavily use it. We have a build with about 200 Modules that basically fall into 3 categories. For each of them we have a parent that defines the tiles to use. Defining the tiles in each module is not really an option in this setup.

And yes, you're right: All tiles are applied at the specified position. This is also the case without this patch, just then the place is fixed to be between the module and the first parent. With this patch, we can control where the tiles are applied and thus can use properties to parameterize some stuff in the tiles in the parent. Again, doing this in each module is not an option.

If 'actual' says to apply the tile before 'parentA' while 'parentA' says to apply it before 'parentB', then AFAICT it would be applied to 'parentA', as the configuration closest to the module wins (it's a "normal" Maven configuration)

Ideally, tiles would be applied before the parent (closest to the module) that included them. But I don't see how this could be implemented, given that we only have a flattened configuration. You could pimp up the configuration to include that information in the <tile>, but then that could still be done later, should there be a need for it.

Patch-set: 3