Update patch set 3
Patch Set 3:

The output is not modified, but it mentions the correct places. If your projects are like this:

me:Module < me:SomeParent < me:UltimateParent

with the tiles to be added before me:SomeParent the output looks like this

[INFO] --- tiles-maven-plugin: Injecting 2 tiles as intermediary parent artifact's...
[INFO] Mixed 'me:SomeParent:1.0.0' with tile 'me.tile:first:1.0.0' as it's new parent.
[INFO] Mixed 'me.tile:first:1.0.0' with tile 'me.tile:second:1.0.0' as it's new parent.
[INFO] Mixed 'me.tile:second:1.0.0' with original parent 'me:UltimateParent:1.0.0' as it's  new top level parent.

I could add output for the project itself, e.g. like this:

[INFO] --- tiles-maven-plugin: Injecting 5 tiles for me:Module:1.0.0 as intermediary parent artifact's...

?

Patch-set: 3
Label: Code-Review=0