Update patch set 4
Patch Set 4:

While the "you don't need parents" works fine for small projects, it does not work for projects where you have several hundred of modules that have different build configurations. The natural thing to do in this case IMHO is to define the tiles in a parent (one parent per different Build config, in our case there are only a handful for ~400 modules). 

Yes, you could push the properties to a tile. But it is much more convenient to define properties in the parent:

- I see the values in the POM
- by using a grandfather POM the values are automatically inherited by everybody. Using a tile you need to take extra care to include it everywhere.

In this case Tiles do not offer any advantage to parents, so I don't see why I should be forced to use them. Just because you've got a hammer, opening a window with it - while working - is not the best way for everybody ;)

I'll add documentation for this, but only if you're willing to merge it. If not, we'll just use a fork and I'll save me the time.

Patch-set: 4