Update patch set 4
Patch Set 4:

Right - so we already existing support that model of operation - sort of.  But by using tiles - the only difference is that the tiles must be in a separate repository to that which consumes them.


pom.xml  (aggregator only)
tile-properties\tile.xml (ultimate parent tile, has all the project's properties)
[typeA]\tile.xml (tile for artifacts of type A. transitively Includes all the tiles, and tile-properties )
[typeB]\tile.xml (tile for artifacts of type B. transitively Includes all the tiles necessary for that, and tile-properties )
..
moduleX\use-tile-type-N (a module, use tile is Type-N)

Which would yield

Module - Type N tile parent - Tile 1 - Tile 2 - Tile Properties...

Any transitively included tile gets added as parents of the current tile set. So the above would define the properties, with the individual tiles above it ( so they get to use the properties, or override them ).

Patch-set: 4
Label: Code-Review=0