m2e LifecycleMapping support in tiles
m2e may need special lifecycle mapping configuration which is supplied
as a "meta" plugin in the <pluginManagement> section. Unfortunately m2e
used project.originalModel, so it will not pick up any configuration
supplied through a tile. In that special case the <pluginManagement>
section of the original model should be updated with the one after tiles
have been applied.

Change-Id: Ifd407c54186936d0371617b6d3841b127de8cc45