m2e LifecycleMapping support in tiles (#9)
m2e may need special lifecycle mapping configuration which is supplied
as a "meta" plugin in the <pluginManagement> section. Unfortunately m2e
uses project.originalModel, so it will not pick up any configuration
supplied through a tile. In that special case the <pluginManagement>
section of the original model should be updated with the one after tiles
have been applied.

To actually work with m2e this still dependes on
https://bugs.eclipse.org/bugs/show_bug.cgi?id=486737 which is not yet
merged.

Change-Id: Ifd407c54186936d0371617b6d3841b127de8cc45