Resolve tiles from current reactor (Fixes #65)
If a tile is defined in the current reactor and used in one of the
modules, it should be resolved directly from the reactor. Otherwise you
need to install the tiles in a pre-build step first.

Also, if resolving from Workspace (e.g. in m2e) make sure the correct
Artifact is used.

Change-Id: I3b30874fabc0833de9be21bde7142dbd7dc5dcf5